fix: cast count() results to INT for backward compatibility
dbt's count() returns BIGINT, which breaks compatibility with external
programs expecting INT (i32). Cast count() results to INT in dbt models
and update schema/entities accordingly.

Changes:
- Cast count() to int in accepted_count.sql, language_count.sql, solver.sql
- Revert schema.sql columns from BIGINT to INT
- Regenerate SeaORM entities with i32
- Update dump-json.rs structs to use i32

+++ dbt/models/marts/language_count.sql
@@ -16,6 +16,6 @@ with simplified_submissions as (
 select
     user_id,
     simplified_language,
-    count(distinct problem_id) as problem_count
+    count(distinct problem_id)::int as problem_count
 from simplified_submissions
 group by user_id, simplified_language
